+++ vite.config.js
@@ -4,6 +4,6 @@ import vue from '@vitejs/plugin-vue'
 export default defineConfig({
   // base: '/viewer-editor/',
   // base: '/mstub-referencer/',
-  base: '/viewer-editor-strapi/',
+  base: '/viewer-hu-pre-and-early-history/',
   plugins: [vue()],
 })